+++ crates/playwright/src/protocol/mod.rs
@@ -106,7 +106,8 @@ pub use mouse::Mouse;
 pub use page::{
     AddLocatorHandlerOptions, AddScriptTagOptions, AddScriptTagOptionsBuilder, AddStyleTagOptions,
     ColorScheme, EmulateMediaOptions, EmulateMediaOptionsBuilder, ForcedColors, GotoOptions, Media,
-    Page, PdfMargin, PdfOptions, PdfOptionsBuilder, ReducedMotion, Response, WaitUntil,
+    Page, PdfMargin, PdfOptions, PdfOptionsBuilder, ReducedMotion, Response, RouteFromHarOptions,
+    WaitUntil,
 };
 pub use playwright::Playwright;
 pub use proxy::ProxySettings;